changed to bigint

+++ db/migrate/20110607071256_add_plan_id_to_subscription_discounts.rb
@@ -1,6 +1,6 @@
 class AddPlanIdToSubscriptionDiscounts < ActiveRecord::Migration
   def self.up
-    add_column :subscription_discounts, :plan_id, :integer
+    add_column :subscription_discounts, :plan_id, :bigint
   end
 
   def self.down
